backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
copilotkit==1.3.0
openai==1.3.0
anthropic==0.18.0
groq==0.4.1
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
supabase==2.3.0
redis==5.0.1
deepagents==0.1.0
